feat(running): 러닝 기록 엔티티 설계 #29

+++ src/main/kotlin/com/runnershi/domain/running/entity/RunningRecord.kt
@@ -8,6 +8,18 @@ import jakarta.persistence.Table
 import java.time.LocalDate
 import java.time.LocalDateTime
 
+// === 단위 기준 ===
+// - distance: m (Int) - 미터 단위 정수 저장, 표시 시 / 1000.0 으로 km 변환
+// - duration: 초 (Int)
+// - pace: 초/km (Int) - 서버에서 distance/duration으로 계산
+// - runningDate: startedAt에서 서버가 추출
+
+// === 정책 ===
+// - pace는 클라이언트에서 받지 않고 서버에서 계산
+// - runningDate는 startedAt 기준으로 서버에서 설정
+// - 유저 탈퇴(soft delete) 시 러닝 기록은 유지
+// - 기록 생성 시 User.totalDistance 동기화 필요
+
 // TODO: 추가 고려 필드
 // 1. 경로 데이터 - GPS 좌표 (별도 테이블 or JSON) - 위치사업자 등록 여부 확인 필요
 // 2. 심박수 - 평균/최대 심박수
@@ -28,15 +40,19 @@ class RunningRecord(
     @Column(name = "user_id", nullable = false)
     val userId: Long,
 
+    // startedAt 기준으로 서버에서 설정
     @Column(name = "running_date", nullable = false)
     val runningDate: LocalDate,
 
+    // 단위: m (미터)
     @Column(nullable = false)
-    val distance: Double,
+    val distance: Int,
 
+    // 단위: 초
     @Column(nullable = false)
     val duration: Int,
 
+    // 단위: 초/km, 서버에서 계산
     @Column(nullable = false)
     val pace: Int,
